+++ ChangeLog.md
@@ -1,8 +1,10 @@
 # ChangeLog
 
-## 1.1.0 (unreleased)
+## 1.1.0 (13 Feb 2022)
 
 - Add `asssembly` branch and adjust [`is-trimmable`](https://github.com/spouliot/cilout/wiki/AssemblyIsTrimmable) command arguments
+- Add [`assembly <assembly> references`](https://github.com/spouliot/cilout/wiki/AssemblyReferences) command
+- Add [`assembly <assembly> definitions`](https://github.com/spouliot/cilout/wiki/AssemblyDefinitions) command
 
 ## 1.0.0 (11 Feb 2022)